Added dto and custom exception

+++ src/service/employee.service.ts
@@ -1,4 +1,6 @@
 import Employee from "../entity/Employee.entity";
+import Address from "../entity/address.entity";
+import HttpException from "../exception/http.exception";
 import EmployeeRepository from "../repository/employee.repository";
 
 class EmployeeService{
@@ -8,29 +10,52 @@ class EmployeeService{
         
     }
 
-    getAllEmployees() : Promise<Employee[]>
+    async getAllEmployees() : Promise<Employee[]>
     {
-        return this.employeeRepository.findAllEmployees();
+        const employees=this.employeeRepository.findAllEmployees();
+
+        if(!employees)
+        {
+            throw new HttpException(404,`Employees not found`);
+        }
+
+        return employees;
     }
 
-    getEmployeeById(id:number):Promise<Employee|null>
+    async getEmployeeById(id:number):Promise<Employee|null>
     {
-        return this.employeeRepository.findAnEmployeeById(id);
+        const employee= await this.employeeRepository.findAnEmployeeById(id);
+        if(!employee)
+        {
+            throw new HttpException(404,`Employee not found with ${id}`);
+        }
+
+        return employee;
     }
 
-    createAnEmployee(name:string,email:string):Promise<Employee>
+    createAnEmployee(name:string,email:string,address:any):Promise<Employee>
     {
         const newEmployee = new Employee();
-        newEmployee.email = name;
-        newEmployee.name = email;
+        newEmployee.email = email;
+        newEmployee.name = name;
+
+        const newAddress=new Address();
+        newAddress.line1=address.line1;
+        newAddress.pincode=address.pincode;
+
+        newEmployee.address=newAddress;
+
+
         return this.employeeRepository.createAnEmployee(newEmployee);
     }
 
-    async updateAnEmployee(id:number,name:string,email:string):Promise<Employee>
+    async updateAnEmployee(id:number,name:string,email:string,address:any):Promise<Employee>
     {
         const employee=await this.employeeRepository.findAnEmployeeById(id);
         employee.name=name;
         employee.email=email;
+        employee.address.line1=address.line1;
+        employee.address.pincode=address.pincode;
         employee.updatedAt=new Date();
         return this.employeeRepository.updateAnEmployee(employee);
     }
@@ -38,6 +63,11 @@ class EmployeeService{
     async deleteEmployee(id:number):Promise<Employee>
     {
         const employee=await this.employeeRepository.findAnEmployeeById(id);
+
+        if(!employee)
+        {
+            throw new HttpException(404,`Employee not found with ${id}`);
+        }
         return this.employeeRepository.deleteEmployee(employee);
     }
